+++ src/browser/modules/D3Visualization/lib/visualization/components/GraphGeometry2.ts
@@ -19,11 +19,11 @@
  */
 
 import GraphStyle from 'browser/modules/D3Visualization/graphStyle'
-import PairwiseArcsRelationshipRouting from '../utils/pairwiseArcsRelationshipRouting'
+import PairwiseArcsRelationshipRouting from '../utils/pairwiseArcsRelationshipRouting2'
 import measureText from '../utils/textMeasurement'
-import Graph from './Graph'
-import Node, { NodeCaptionLine } from './Node'
-import Relationship from './Relationship'
+import Graph from './Graph2'
+import Node, { NodeCaptionLine } from './Node2'
+import Relationship from './relationship2'
 
 export default class GraphGeometry {
   relationshipRouting: PairwiseArcsRelationshipRouting
