+++ src/browser/modules/D3Visualization/lib/visualization/renders/init.ts
@@ -17,8 +17,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import Node, { NodeCaptionLine } from '../components/Node'
-import Renderer from '../components/Renderer'
+import Node, { NodeCaptionLine } from '../components/Node2'
+import Renderer from '../components/renderer2'
 const noop = () => undefined
 
 const nodeRingStrokeSize = 8
