+++ src/browser/modules/D3Visualization/lib/visualization/utils/LoopArrow2.ts
