+++ src/browser/modules/D3Visualization/lib/visualization/utils/StraightArrow2.ts
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { RelationShipCaptionLayout } from '../components/Relationship'
+import { RelationShipCaptionLayout } from '../components/relationship2'
 
 export default class StraightArrow {
   length: number
